Restructring to have a better plan for the future commands
