+++ lib/movement/commands.rb
@@ -1,7 +1,9 @@
-require "movement/generate_command"
+require "movement/commands/generate"
 
 module Movement
-  class Commands < Thor
-    register(Movement::GenerateCommand, "generate", "generate", "Generate resources for your app")
+  module Commands
+    class Core < Thor
+      register(Commands::Generate, "generate", "generate", "Generate resources for your app")
+    end
   end
 end
