+++ lib/movement/commands/generate.rb
@@ -1,0 +1,20 @@
+module Movement
+  module Commands
+    class Generate < Thor::Group
+      include Thor::Actions
+
+      argument :generator
+      argument :name
+
+      def self.source_root
+        File.dirname(__FILE__)
+      end
+
+      def generate
+        if generator == "view"
+          template("templates/UIView.tt", "app/views/#{name}.rb")
+        end
+      end
+    end
+  end
+end
