+++ lib/movement/commands/templates/UIView.tt
