+++ movement.gemspec
@@ -21,4 +21,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "bundler", "~> 1.7"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_dependency "thor", "~> 0.19"
+  spec.add_dependency "activesupport", "~> 4.1"
 end
